# api/openapi.yaml
openapi: "3.1.0"
info:
  title: NovaFabric Multi-User REST API
  version: "0.7.0"
  description: |
    Multi-user REST API for the NovaFabric AI Asset Registry.
    URL prefix: /v0/
    Authentication: Bearer token via Authorization header (OIDC) or open when OIDC disabled.
    All list endpoints use cursor-based pagination (limit + next_cursor).
    All error responses use the standard error envelope.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /v0
    description: NovaFabric server (versioned)

# --------------------------------------------------------------------------- #
# Reusable components
# --------------------------------------------------------------------------- #
components:
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Asset 'my-agent@v1.0' not found.
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"

    # -------- pagination --------
...
    EvidenceExportRequest:
      type: object
      required: [run_id]
      properties:
        run_id:
          type: string
          description: The capsule run_id to export as evidence.
        output_path:
          type: string
          nullable: true
          description: Optional override for the output ZIP path (server-side path).
        allow_unsafe_skips:
          type: boolean
          default: false
    BundleSummary:
      type: object
      required: [bundle_id, run_id, created_at, size_bytes]
      properties:
        bundle_id:
          type: string
        run_id:
          type: string
        created_at:
          type: string
          format: date-time
        size_bytes:
          type: integer
        bundle_path:
          type: string
          nullable: true
...
  responses:
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Resource conflict (e.g. duplicate asset or invalid lifecycle transition).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: Spec validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

# --------------------------------------------------------------------------- #
# Paths
# --------------------------------------------------------------------------- #
paths:
...
  /evidence:
    post:
      operationId: exportEvidence
      summary: Build an evidence bundle ZIP from a capsule
      tags: [evidence]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceExportRequest"
      responses:
        "202":
          description: Evidence export accepted; poll GET /evidence/{bundle_id}.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /evidence/{bundle_id}:
    get:
      operationId: getEvidenceBundle
      summary: Get evidence bundle metadata
      tags: [evidence]
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bundle metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleSummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /evidence/{bundle_id}/download:
    get:
      operationId: downloadEvidenceBundle
      summary: Download an evidence bundle ZIP
      tags: [evidence]
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ZIP file binary download.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------------------------------------------------- #
  # Auth — Device Authorization Grant (RFC 8628)
  # --------------------------------------------------------------------------- #
  /auth/device/code:
    post:
      operationId: issueDeviceCode
      summary: Issue a device code (Device Authorization Grant)
      tags: [auth]
      description: |
        Initiates the Device Authorization Grant flow (RFC 8628).
        Returns a device_code and user_code; the client polls /auth/token
        until the user approves at the verification_uri.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                server_url:
                  type: string
                  example: "http://localhost:7433"
      responses:
        "200":
          description: Device code issued.
          content:
            application/json:
              schema:
                type: object
                required: [device_code, user_code, verification_uri, expires_in, interval]
                properties:
                  device_code:
                    type: string
                  user_code:
                    type: string
                    example: "ABCD-1234"
                  verification_uri:
                    type: string
                  expires_in:
                    type: integer
                    description: Seconds until the device code expires.
                  interval:
                    type: integer
                    description: Minimum polling interval in seconds.
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/token:
    post:
      operationId: pollDeviceToken
      summary: Poll for a token after device code approval
      tags: [auth]
      description: |
        Client polls this endpoint until the user approves the device code.
        Returns the access token on approval, or a pending/expired error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_code]
              properties:
                device_code:
                  type: string
      responses:
        "200":
          description: Access token issued.
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type]
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
        "400":
          description: Pending approval or expired device code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/approve:
    post:
      operationId: approveDeviceCode
      summary: Approve a pending device code (admin / testing)
      tags: [auth]
      description: |
        Admin endpoint that approves a pending device code, associating a
        subject and roles with it. Used for testing or headless admin flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_code, subject]
              properties:
                user_code:
                  type: string
                subject:
                  type: string
                  example: "user@example.com"
                roles:
                  type: array
                  items:
                    type: string
                  example: ["reader"]
      responses:
        "200":
          description: Device code approved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------------------------------------------------- #
  # Admin
  # --------------------------------------------------------------------------- #
...